requirements.txt
fastapi
uvicorn
requests
cortex4py
pydantic
python-dotenv